[update] 代码优化

+++ src/main/java/com/aicallsystem/acs/common/BaseEntity.java
@@ -3,11 +3,10 @@ package com.aicallsystem.acs.common;
 import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 /**
  * <p>
- *   对象基础配置
+ *      entity基础配置
  * </p>
  * @author ispong
  * @since 4/3/2019
@@ -19,6 +18,5 @@ public class BaseEntity {
     private static final long serialVersionUID = 1L;
 
     @TableId
-    private BigDecimal id;
-
+    private int id;
 }
